# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SIP",
        "description": "Sistem informasi penjualan",
        "contact": {
            "email": "admin@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Url from config file"
        },
        {
            "url": "https://google.com",
            "description": "Production url"
        }
    ],
    "paths": {
...
        "/api/product/all": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get all product",
                "description": "get all product",
                "operationId": "get-all-product",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/product/{product_id}": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "Get detail product",
                "description": "get detail product",
                "operationId": "get-detail-product",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Product"
                ],
                "summary": "Delete product",
                "description": "delete product",
                "operationId": "delete-product",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Data not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/product": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Create product",
                "description": "Create product",
                "operationId": "create-product",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Category id",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "integer"
                                    },
                                    "sku": {
                                        "description": "SKU",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "image": {
                                        "description": "Image",
                                        "type": "file"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/product/update/{product_id}": {
            "post": {
                "tags": [
                    "Product"
                ],
                "summary": "Update product",
                "description": "update product",
                "operationId": "update-product",
                "parameters": [
                    {
                        "name": "product_id",
                        "in": "path"
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "category_id": {
                                        "description": "Category id",
                                        "type": "integer"
                                    },
                                    "name": {
                                        "description": "Name",
                                        "type": "string"
                                    },
                                    "price": {
                                        "description": "Price",
                                        "type": "integer"
                                    },
                                    "sku": {
                                        "description": "SKU",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive"
                                        ]
                                    },
                                    "image": {
                                        "description": "Image",
                                        "type": "file"
                                    },
                                    "description": {
                                        "description": "Description",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "api_key": {
                "type": "apiKey",
                "description": "A short description for security scheme",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        []
    ]
}